refactor(testing): modularize TESTING sub-agent from 1097 LOC to <550 LOC per file
Refactors lib/sub-agents/testing.js (1097 LOC) into modular structure:
- lib/sub-agents/testing/index.js (544 LOC) - Main orchestrator
- lib/sub-agents/testing/phases/phase1-preflight.js (88 LOC)
- lib/sub-agents/testing/phases/phase2-generation.js (54 LOC)
- lib/sub-agents/testing/phases/phase3-execution.js (147 LOC)
- lib/sub-agents/testing/phases/phase4-evidence.js (111 LOC)
- lib/sub-agents/testing/phases/phase5-verdict.js (161 LOC)
- lib/sub-agents/testing/utils/troubleshooting.js (137 LOC)

Original testing.js now re-exports from testing/index.js for backward
compatibility with the sub-agent executor import mechanism.

SD: SD-LEO-REFAC-TESTING-INFRA-001
