+++ lib/sub-agents/testing.js
@@ -1,5 +1,5 @@
 /**
- * TESTING Sub-Agent (QA Engineering Director v3.0 - Intelligence Enhanced)
+ * TESTING Sub-Agent (QA Engineering Director v3.1 - Modular Architecture)
  * LEO Protocol v4.2.0 - Sub-Agent Performance Enhancement
  *
  * Purpose: Mission-Critical Testing Automation - Comprehensive E2E validation
@@ -8,1090 +8,24 @@
  *
  * Philosophy: "Do it right, not fast." E2E testing is MANDATORY, not optional.
  *
- * v3.0 Enhancements (2025-11-21):
- * - Phase 0: Intelligent Test Analysis
- *   - Selector validation: Auto-detect mismatches before running tests
- *   - Navigation flow validation: Verify navigation paths exist
- *   - Component mapping: Ensure test-to-component alignment
- *   - Contextual error analysis: Smart fix suggestions
+ * REFACTORED: v3.1 (2026-01-23)
+ * This file now re-exports from modular structure for maintainability.
+ * See lib/sub-agents/testing/ directory for implementation:
+ *   - index.js - Main orchestrator (~550 LOC)
+ *   - phases/phase1-preflight.js - Pre-flight checks
+ *   - phases/phase2-generation.js - Test case generation
+ *   - phases/phase3-execution.js - E2E test execution
+ *   - phases/phase4-evidence.js - Evidence collection
+ *   - phases/phase5-verdict.js - Verdict generation
+ *   - utils/troubleshooting.js - Troubleshooting tactics
  *
- * Created: 2025-10-11 (SD-SUBAGENT-IMPROVE-001)
+ * SD: SD-LEO-REFAC-TESTING-INFRA-001
+ *
+ * Original: 2025-10-11 (SD-SUBAGENT-IMPROVE-001)
  * Updated: 2025-11-15 (SD-LEO-PROTOCOL-V4-4-0: Adaptive validation support)
  * Updated: 2025-11-21 (v3.0: Phase 1 Intelligence Module - SD-FOUND-DATA-003)
+ * Refactored: 2026-01-23 (v3.1: Modular architecture - SD-LEO-REFAC-TESTING-INFRA-001)
  */
 
-import path from 'path';
-import { fileURLToPath } from 'url';
-import dotenv from 'dotenv';
-import {
-  detectValidationMode,
-  logValidationMode
-} from '../utils/adaptive-validation.js';
-import {
-  validateTestSelectors,
-  validateNavigationFlow,
-  analyzeTestComponentMapping
-} from '../utils/test-intelligence.js';
-import { createSupabaseServiceClient } from '../../scripts/lib/supabase-connection.js';
-// LEO v4.3.4: Unified test evidence architecture
-import {
-  checkTestEvidenceFreshness
-} from '../../scripts/lib/test-evidence-ingest.js';
-// TIER 1.5: Handoff preflight check
-import { quickPreflightCheck } from '../../scripts/lib/handoff-preflight.js';
-// LEO v4.4.3: Branch-aware test scanning for feature branch validation
-import { resolveBranch } from '../../scripts/lib/branch-resolver.js';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-dotenv.config();
-// Supabase client initialized in execute() to use async createSupabaseServiceClient
-let supabase = null;
-
-/**
- * Execute TESTING sub-agent
- * Implements QA Engineering Director v2.0 workflow
- *
- * @param {string} sdId - Strategic Directive ID
- * @param {Object} subAgent - Sub-agent instructions (already loaded)
- * @param {Object} options - Execution options
- * @returns {Promise<Object>} Testing results
- */
-export async function execute(sdId, subAgent, options = {}) {
-  console.log(`\n🧪 Starting TESTING for ${sdId}...`);
-  console.log('   QA Engineering Director v2.0 - Testing-First Edition');
-
-  // Initialize Supabase client with SERVICE_ROLE_KEY (SD-FOUND-SAFETY-002 pattern)
-  if (!supabase) {
-    supabase = await createSupabaseServiceClient('engineer', { verbose: false });
-  }
-
-  // TIER 1.5: Handoff Preflight Check
-  // Verify SD has proper handoff chain before proceeding with testing validation
-  // This is advisory - we log warnings but don't block testing work
-  try {
-    console.log('   🔗 Checking handoff chain status...');
-    const preflightResult = await quickPreflightCheck(sdId, 'EXEC');
-
-    if (preflightResult.ready) {
-      console.log('   ✅ Handoff chain verified for EXEC phase');
-    } else {
-      console.log('   ⚠️  Handoff chain incomplete:');
-      (preflightResult.missing || []).forEach(h => console.log(`      • Missing: ${h}`));
-      console.log('   💡 Consider running: node scripts/handoff.js create --sd ' + sdId);
-      console.log('   ⚠️  Proceeding with TESTING validation (advisory check)');
-    }
-  } catch (preflightError) {
-    console.log(`   ⚠️  Handoff preflight skipped: ${preflightError.message}`);
-  }
-
-  // LEO v4.4.3: Resolve feature branch for branch-aware test scanning
-  // Tests live on feature branch until SD is approved and merged
-  let branchContext = null;
-  try {
-    console.log('   🌿 Resolving feature branch...');
-    const branchResult = await resolveBranch(supabase, sdId, {
-      verbose: false,
-      autoStore: true
-    });
-
-    if (branchResult.success) {
-      branchContext = {
-        branch: branchResult.branch,
-        repoPath: branchResult.repoPath,
-        source: branchResult.source,
-        validated: branchResult.validated
-      };
-      console.log(`   ✅ Feature branch resolved: ${branchResult.branch}`);
-      console.log(`      Source: ${branchResult.source}, Repo: ${branchResult.repoPath}`);
-    } else {
-      console.log(`   ⚠️  Could not resolve feature branch: ${branchResult.error}`);
-      console.log('   💡 Falling back to filesystem scan (may miss tests on feature branch)');
-    }
-  } catch (branchError) {
-    console.log(`   ⚠️  Branch resolution error: ${branchError.message}`);
-  }
-
-  // Fetch semantic patterns from user stories' e2e_test_path
-  let semanticPatterns = [];
-  try {
-    const { data: stories } = await supabase
-      .from('user_stories')
-      .select('e2e_test_path')
-      .or(`sd_id.eq.${sdId},sd_id.ilike.%${sdId}%`)
-      .not('e2e_test_path', 'is', null);
-
-    if (stories && stories.length > 0) {
-      // Extract unique filename patterns from e2e_test_path
-      const paths = stories.map(s => s.e2e_test_path).filter(Boolean);
-      const uniqueFilenames = [...new Set(paths.map(p => {
-        // Extract filename without extension, convert to pattern
-        const match = p.match(/([^/]+)\.spec\.ts$/);
-        return match ? match[1].toLowerCase() : null;
-      }).filter(Boolean))];
-      semanticPatterns = uniqueFilenames;
-      console.log(`   📝 Semantic patterns from user stories: ${uniqueFilenames.join(', ')}`);
-    }
-  } catch (e) {
-    console.log(`   ⚠️  Could not fetch semantic patterns: ${e.message}`);
-  }
-
-  // Build enhanced options with branch context
-  const enhancedOptions = {
-    ...options,
-    ...(branchContext && {
-      branch: branchContext.branch,
-      featureBranch: branchContext.branch,
-      repoPath: branchContext.repoPath
-    }),
-    semanticPatterns
-  };
-
-  // SD-LEO-PROTOCOL-V4-4-0: Detect validation mode (prospective vs retrospective)
-  const validationMode = await detectValidationMode(sdId, options);
-
-  logValidationMode('TESTING', validationMode, {
-    'Prospective': 'BLOCKED if --full-e2e flag missing',
-    'Retrospective': 'CONDITIONAL_PASS if E2E tests exist and pass'
-  });
-
-  // PAT-DB-SD-E2E-001: Database/Infrastructure SDs don't require UI E2E tests
-  // Check SD type and return early PASS for non-UI SDs
-  const { data: sdData } = await supabase
-    .from('strategic_directives_v2')
-    .select('category')
-    .or(`legacy_id.eq.${sdId},id.eq.${sdId}`)
-    .single();
-
-  const sdCategory = sdData?.category?.toLowerCase() || '';
-  // PAT-DB-SD-E2E-001: Skip UI E2E for non-UI SDs
-  // 'refactor' added per SD-REFACTOR-SCRIPTS-001: structural refactors use REGRESSION, not E2E
-  const skipE2ESdTypes = ['database', 'infrastructure', 'documentation', 'protocol', 'refactor'];
-
-  if (skipE2ESdTypes.includes(sdCategory)) {
-    const isRefactor = sdCategory === 'refactor';
-    console.log(`\n🗄️  SD Type Detection: ${sdCategory.toUpperCase()}`);
-    console.log(isRefactor
-      ? '   💡 Refactor SDs use REGRESSION sub-agent for validation'
-      : '   💡 Database/Infrastructure SDs do not require UI E2E tests');
-    console.log(isRefactor
-      ? '   ✅ Validation via REGRESSION sub-agent (before/after behavior comparison)'
-      : '   ✅ Validation via DATABASE/SECURITY sub-agents + table existence');
-
-    return {
-      verdict: 'PASS',
-      confidence: 95,
-      validation_mode: validationMode,
-      critical_issues: [],
-      warnings: [],
-      recommendations: [{
-        severity: 'INFO',
-        issue: `${sdCategory} SD - UI E2E tests not applicable`,
-        recommendation: isRefactor
-          ? 'Behavior validated via REGRESSION sub-agent (before/after comparison)'
-          : 'Schema validated via DATABASE sub-agent and table existence checks'
-      }],
-      detailed_analysis: {
-        sd_type: sdCategory,
-        skip_reason: isRefactor
-          ? 'Refactor SD - validation via REGRESSION sub-agent behavior comparison'
-          : 'Non-UI SD type - E2E validation deferred to DATABASE/SECURITY sub-agents',
-        validation_approach: isRefactor
-          ? 'REGRESSION sub-agent: before/after behavior comparison, no functional changes'
-          : 'SQL schema validation, RLS policy verification'
-      },
-      findings: {
-        phase0_intelligence: { skipped: true, reason: `${sdCategory} SD - no UI components` },
-        phase1_preflight: { skipped: true },
-        phase2_test_generation: { skipped: true },
-        phase3_execution: { skipped: true, reason: 'E2E not applicable for database SDs' },
-        phase4_evidence: { type: 'schema_validation' },
-        phase5_verdict: { auto_pass: true, sd_type: sdCategory }
-      },
-      options
-    };
-  }
-
-  const results = {
-    verdict: 'PASS',
-    confidence: 100,
-    validation_mode: validationMode,  // Add validation mode to results
-    critical_issues: [],
-    warnings: [],
-    recommendations: [],
-    detailed_analysis: {},
-    findings: {
-      phase0_intelligence: null,  // NEW: Intelligent pre-flight analysis
-      phase1_preflight: null,
-      phase2_test_generation: null,
-      phase3_execution: null,
-      phase4_evidence: null,
-      phase5_verdict: null
-    },
-    options
-  };
-
-  try {
-    // Phase 0: Intelligent Test Analysis (NEW)
-    console.log('\n🧠 Phase 0: Intelligent Test Analysis (v3.0 Enhanced)...');
-    const phase0 = {
-      selector_validation: await validateTestSelectors(sdId, enhancedOptions),
-      navigation_validation: await validateNavigationFlow(sdId, enhancedOptions),
-      component_mapping: await analyzeTestComponentMapping(sdId, enhancedOptions)
-    };
-    results.findings.phase0_intelligence = phase0;
-
-    // Check for critical issues found by intelligence module
-    if (phase0.selector_validation.mismatches_found > 0) {
-      const mismatchCount = phase0.selector_validation.mismatches_found;
-      results.warnings.push({
-        severity: 'HIGH',
-        issue: `Found ${mismatchCount} selector mismatch(es) in test files`,
-        recommendation: 'Review Phase 0 suggestions and fix selectors before running tests',
-        details: phase0.selector_validation.suggestions.slice(0, 3),
-        confidence: phase0.selector_validation.confidence
-      });
-
-      // Reduce confidence based on severity
-      if (phase0.selector_validation.confidence < 50) {
-        results.confidence = Math.min(results.confidence, 60);
-      }
-    }
-
-    if (phase0.navigation_validation.broken_paths.length > 0) {
-      results.warnings.push({
-        severity: 'HIGH',
-        issue: `Found ${phase0.navigation_validation.broken_paths.length} broken navigation path(s)`,
-        recommendation: 'Fix navigation sequences before running E2E tests',
-        details: phase0.navigation_validation.broken_paths
-      });
-    }
-
-    if (phase0.component_mapping.missing_components.length > 0) {
-      results.warnings.push({
-        severity: 'MEDIUM',
-        issue: `${phase0.component_mapping.missing_components.length} component(s) referenced in tests not found`,
-        recommendation: 'Verify component paths or remove invalid test references',
-        details: phase0.component_mapping.missing_components
-      });
-    }
-
-    // Phase 1: Pre-flight Checks
-    console.log('\n🔍 Phase 1: Pre-flight Checks...');
-    const phase1 = await preflightChecks(sdId, options);
-    results.findings.phase1_preflight = phase1;
-
-    if (phase1.blocked) {
-      results.verdict = 'BLOCKED';
-      results.confidence = 100;
-      results.critical_issues.push(...phase1.critical_issues);
-      return results;
-    }
-
-    if (phase1.warnings.length > 0) {
-      results.warnings.push(...phase1.warnings);
-      if (results.confidence > 90) results.confidence = 90;
-    }
-
-    // Phase 2: Professional Test Case Generation
-    console.log('\n📝 Phase 2: Test Case Generation...');
-    const phase2 = await generateTestCases(sdId, options);
-    results.findings.phase2_test_generation = phase2;
-
-    if (phase2.user_stories_count === 0) {
-      console.log('   ⚠️  No user stories found - cannot generate test cases');
-      results.warnings.push({
-        severity: 'HIGH',
-        issue: 'No user stories found for SD',
-        recommendation: 'Create user stories before testing (Product Requirements Expert)',
-        note: 'E2E tests should map to user stories (100% coverage required)'
-      });
-      if (results.confidence > 70) results.confidence = 70;
-    }
-
-    // Phase 2.5: Check for fresh test evidence (LEO v4.3.4)
-    // If we have fresh evidence and we're in retrospective mode, we can skip re-running tests
-    console.log('\n🔍 Phase 2.5: Checking for existing test evidence...');
-    let skipTestExecution = false;
-    let freshEvidence = null;
-
-    try {
-      const freshnessCheck = await checkTestEvidenceFreshness(sdId, 60); // 60 minute threshold
-      results.findings.phase2_5_evidence_check = freshnessCheck;
-
-      if (freshnessCheck.isFresh && validationMode === 'retrospective') {
-        console.log(`   ✅ Found FRESH test evidence (${Math.round(freshnessCheck.ageMinutes || 0)} minutes old)`);
-        console.log(`   📊 Previous verdict: ${freshnessCheck.evidence?.verdict}, Pass rate: ${freshnessCheck.evidence?.pass_rate}%`);
-
-        // Use existing evidence instead of re-running tests
-        if (freshnessCheck.evidence?.verdict === 'PASS' || freshnessCheck.evidence?.pass_rate >= 95) {
-          skipTestExecution = true;
-          freshEvidence = freshnessCheck.evidence;
-          console.log('   ⏭️  Skipping test execution - using existing evidence (retrospective mode)');
-        }
-      } else if (freshnessCheck.evidence) {
-        console.log(`   ⚠️  Test evidence is ${freshnessCheck.freshnessStatus} (${Math.round(freshnessCheck.ageMinutes || 0)} minutes old)`);
-        console.log(`   ${freshnessCheck.recommendation || 'Consider re-running tests'}`);
-      } else {
-        console.log('   📭 No E2E test evidence found in test_runs');
-      }
-    } catch (evidenceError) {
-      console.log(`   ⚠️  Could not check E2E test evidence: ${evidenceError.message}`);
-    }
-
-    // LEO v4.4.1: Fallback check for API integration tests in sd_testing_status
-    // API/backend SDs may have integration tests instead of E2E tests
-    if (!skipTestExecution) {
-      try {
-        console.log('   🔍 Checking for API integration test evidence (sd_testing_status)...');
-        const { data: apiTestEvidence, error: apiTestError } = await supabase
-          .from('sd_testing_status')
-          .select('*')
-          .eq('sd_id', sdId)
-          .single();
-
-        if (!apiTestError && apiTestEvidence && apiTestEvidence.tested) {
-          const passRate = apiTestEvidence.test_pass_rate ||
-            (apiTestEvidence.test_count > 0
-              ? (apiTestEvidence.tests_passed / apiTestEvidence.test_count) * 100
-              : 0);
-          const ageMinutes = apiTestEvidence.last_tested_at
-            ? Math.round((Date.now() - new Date(apiTestEvidence.last_tested_at).getTime()) / 60000)
-            : null;
-          const isFresh = ageMinutes !== null && ageMinutes <= 60;
-
-          console.log('   ✅ Found API integration test evidence');
-          console.log(`      Framework: ${apiTestEvidence.test_framework || 'unknown'}`);
-          console.log(`      Tests: ${apiTestEvidence.tests_passed}/${apiTestEvidence.test_count} passed (${passRate.toFixed(1)}%)`);
-          console.log(`      Age: ${ageMinutes !== null ? ageMinutes + ' minutes' : 'unknown'}`);
-
-          if (isFresh && passRate >= 95) {
-            skipTestExecution = true;
-            freshEvidence = {
-              verdict: passRate >= 100 ? 'PASS' : 'CONDITIONAL_PASS',
-              pass_rate: passRate,
-              total_tests: apiTestEvidence.test_count,
-              passed_tests: apiTestEvidence.tests_passed,
-              failed_tests: apiTestEvidence.tests_failed,
-              duration_ms: (apiTestEvidence.test_duration_seconds || 0) * 1000,
-              age_minutes: ageMinutes,
-              test_type: 'api_integration',
-              framework: apiTestEvidence.test_framework
-            };
-            results.findings.phase2_5_api_test_evidence = freshEvidence;
-            console.log('   ⏭️  Using API integration test evidence (pass rate meets threshold)');
-          } else if (!isFresh) {
-            console.log('   ⚠️  API test evidence is stale (> 60 minutes)');
-          } else {
-            console.log(`   ⚠️  API test pass rate ${passRate.toFixed(1)}% below threshold (95%)`);
-          }
-        } else if (apiTestError?.code !== 'PGRST116') {
-          // PGRST116 = no rows found (expected for SDs without API tests)
-          console.log(`   ⚠️  Could not check API test evidence: ${apiTestError?.message || 'unknown error'}`);
-        } else {
-          console.log('   📭 No API integration test evidence found');
-        }
-      } catch (apiCheckError) {
-        console.log(`   ⚠️  API test evidence check failed: ${apiCheckError.message}`);
-      }
-    }
-
-    if (!skipTestExecution) {
-      console.log('   📋 Proceeding with standard test execution');
-    }
-
-    // Phase 3: E2E Test Execution (MANDATORY - unless we have fresh evidence)
-    console.log('\n🚀 Phase 3: E2E Test Execution (MANDATORY)...');
-    let phase3;
-
-    if (skipTestExecution && freshEvidence) {
-      // Use existing evidence instead of running tests
-      phase3 = {
-        tests_executed: freshEvidence.total_tests,
-        tests_passed: freshEvidence.passed_tests,
-        failed_tests: freshEvidence.failed_tests,
-        failures: [],
-        report_url: freshEvidence.report_file_path,
-        execution_time_ms: freshEvidence.duration_ms,
-        evidence_reused: true,
-        evidence_age_minutes: Math.round(freshEvidence.age_minutes || 0)
-      };
-      console.log(`   📋 Using cached evidence: ${phase3.tests_passed}/${phase3.tests_executed} tests passed`);
-    } else {
-      // Run actual E2E tests
-      phase3 = await executeE2ETests(sdId, options);
-    }
-    results.findings.phase3_execution = phase3;
-
-    if (phase3.failed_tests > 0) {
-      // Check pass rate threshold before marking as critical
-      const passRate = phase3.tests_executed > 0 ? (phase3.tests_passed / phase3.tests_executed) * 100 : 0;
-      const PASS_RATE_THRESHOLD = 95;
-
-      if (passRate >= PASS_RATE_THRESHOLD) {
-        console.log(`   ⚠️  ${phase3.failed_tests} test(s) failed, but pass rate ${passRate.toFixed(1)}% meets threshold (≥${PASS_RATE_THRESHOLD}%)`);
-        // Add as warning, not critical issue
-        results.warnings.push({
-          severity: 'MEDIUM',
-          issue: `${phase3.failed_tests} E2E test(s) failed`,
-          recommendation: 'Consider fixing for 100% test coverage',
-          details: phase3.failures,
-          pass_rate: passRate
-        });
-      } else {
-        console.log(`   ❌ ${phase3.failed_tests} test(s) failed (pass rate ${passRate.toFixed(1)}% below threshold)`);
-        results.critical_issues.push({
-          severity: 'CRITICAL',
-          issue: `${phase3.failed_tests} E2E test(s) failed (pass rate ${passRate.toFixed(1)}% < ${PASS_RATE_THRESHOLD}%)`,
-          recommendation: 'Fix test failures before proceeding',
-          details: phase3.failures,
-          pass_rate: passRate
-        });
-        results.verdict = 'BLOCKED';
-      }
-    } else if (phase3.tests_executed === 0) {
-      console.log('   ⚠️  No E2E tests executed');
-      results.warnings.push({
-        severity: 'HIGH',
-        issue: 'No E2E tests executed',
-        recommendation: 'Create and execute E2E tests (MANDATORY per protocol)',
-        note: 'Cannot approve SD without E2E test evidence'
-      });
-      if (results.confidence > 60) results.confidence = 60;
-    } else {
-      console.log(`   ✅ All ${phase3.tests_passed} test(s) passed`);
-    }
-
-    // Phase 4: Evidence Collection
-    console.log('\n📸 Phase 4: Evidence Collection...');
-    const phase4 = await collectEvidence(sdId, phase3);
-    results.findings.phase4_evidence = phase4;
-
-    // Phase 4.5: User Story Verification (Quick-Fix QF-20251206-001)
-    // Updated: Accept validation_status='validated' for backend/non-UI SDs (per migration 20251016)
-    console.log('\n📋 Phase 4.5: User Story Verification...');
-    const { data: stories, error: storyError } = await supabase
-      .from('user_stories')
-      .select('story_key, title, status, e2e_test_path, e2e_test_status, validation_status')
-      .eq('sd_id', sdId);
-
-    if (!storyError && stories && stories.length > 0) {
-      // A story is complete if:
-      // 1. status is 'completed' or 'validated'
-      // 2. AND (e2e_test_status = 'passing' OR validation_status = 'validated')
-      // This allows backend SDs with 'skipped' E2E but 'validated' validation_status
-      const incomplete = stories.filter(s =>
-        !['completed', 'validated'].includes(s.status) ||
-        !s.e2e_test_path ||
-        (s.e2e_test_status !== 'passing' && s.validation_status !== 'validated')
-      );
-
-      if (incomplete.length > 0) {
-        console.log(`   ❌ ${incomplete.length} user stories not fully implemented`);
-        incomplete.forEach(s => {
-          console.log(`      - ${s.story_key}: ${s.title}`);
-          console.log(`        Status: ${s.status || 'NULL'}, E2E: ${s.e2e_test_path || 'NOT MAPPED'}, Result: ${s.e2e_test_status || 'NOT RUN'}, Validation: ${s.validation_status || 'pending'}`);
-        });
-        results.critical_issues.push({
-          severity: 'CRITICAL',
-          issue: `${incomplete.length} user stories not fully implemented`,
-          stories: incomplete.map(s => ({ story_key: s.story_key, status: s.status, e2e_mapped: !!s.e2e_test_path })),
-          recommendation: 'Complete implementation of all user stories before EXEC→PLAN handoff'
-        });
-        results.verdict = 'BLOCKED';
-        results.confidence = Math.max(0, 100 - (incomplete.length / stories.length * 100));
-      } else {
-        console.log(`   ✅ All ${stories.length} user stories fully implemented`);
-      }
-    } else if (storyError) {
-      console.log('   ⚠️  Could not verify user stories:', storyError.message);
-      results.warnings.push('User story verification failed - check manually');
-    }
-
-    // Phase 5: Verdict & Testing Learnings
-    console.log('\n🏁 Phase 5: Verdict & Testing Learnings...');
-    const phase5 = generateVerdict(results, validationMode);  // Pass validation mode
-    results.findings.phase5_verdict = phase5;
-
-    results.verdict = phase5.verdict;
-    results.confidence = phase5.confidence;
-    results.recommendations = phase5.recommendations;
-
-    // SD-LEO-PROTOCOL-V4-4-0: Add justification and conditions for CONDITIONAL_PASS
-    if (phase5.justification) results.justification = phase5.justification;
-    if (phase5.conditions) results.conditions = phase5.conditions;
-
-    console.log(`\n✅ TESTING Complete: ${results.verdict} (${results.confidence}% confidence)`);
-
-    return results;
-
-  } catch (error) {
-    console.error('\n❌ TESTING error:', error.message);
-    results.verdict = 'FAIL';
-    results.error = error.message;
-    results.confidence = 0;
-    results.critical_issues.push({
-      severity: 'CRITICAL',
-      issue: 'TESTING sub-agent execution failed',
-      recommendation: 'Review error and retry',
-      error: error.message
-    });
-    return results;
-  }
-}
-
-/**
- * Phase 1: Pre-flight Checks
- */
-async function preflightChecks(sdId, options) {
-  const checks = {
-    blocked: false,
-    critical_issues: [],
-    warnings: [],
-    build_status: null,
-    migrations_status: null,
-    component_integration: null
-  };
-
-  // Check 1: Build validation (if not skipped)
-  if (!options.skip_build) {
-    console.log('   🏗️  Checking build status...');
-    try {
-      // This is a simplified check - in reality, would run actual build
-      checks.build_status = {
-        passed: true,
-        message: 'Build check skipped (would run: npm run build)'
-      };
-      console.log('      ✅ Build validation passed');
-    } catch (error) {
-      checks.blocked = true;
-      checks.critical_issues.push({
-        severity: 'CRITICAL',
-        issue: 'Build failed',
-        recommendation: 'Fix build errors before testing',
-        error: error.message
-      });
-      console.log('      ❌ Build validation failed');
-    }
-  } else {
-    console.log('   ⏭️  Build validation skipped');
-    checks.build_status = { skipped: true };
-  }
-
-  // Check 2: Database migration verification
-  console.log('   🗄️  Checking database migrations...');
-  try {
-    const { data: migrations, error } = await supabase
-      .from('migrations')
-      .select('*')
-      .order('created_at', { ascending: false })
-      .limit(5);
-
-    if (error) {
-      console.log(`      ⚠️  Could not check migrations: ${error.message}`);
-      checks.warnings.push({
-        severity: 'MEDIUM',
-        issue: 'Could not verify database migrations',
-        recommendation: 'Manually verify migrations are applied'
-      });
-    } else {
-      console.log(`      ✅ Migration check complete (${migrations?.length || 0} recent)`);
-      checks.migrations_status = {
-        checked: true,
-        recent_count: migrations?.length || 0
-      };
-    }
-  } catch (error) {
-    console.log(`      ⚠️  Migration check error: ${error.message}`);
-  }
-
-  // Check 3: Component integration (manual note)
-  console.log('   🔗 Component integration check...');
-  console.log('      💡 Tip: Verify components are imported and used (not just created)');
-  checks.component_integration = {
-    manual_check_required: true,
-    suggestion: 'Search for component imports in parent files'
-  };
-
-  return checks;
-}
-
-/**
- * Phase 2: Generate Test Cases from User Stories
- */
-async function generateTestCases(sdId, _options) {
-  console.log('   📋 Querying user stories...');
-
-  const { data: userStories, error } = await supabase
-    .from('user_stories')
-    .select('*')
-    .eq('sd_id', sdId)
-    .order('story_points', { ascending: false });
-
-  if (error) {
-    console.log(`      ⚠️  Could not query user stories: ${error.message}`);
-    return {
-      user_stories_count: 0,
-      error: error.message
-    };
-  }
-
-  const count = userStories?.length || 0;
-  console.log(`      ✅ Found ${count} user stories`);
-
-  if (count > 0) {
-    console.log(`      💡 Test Coverage Target: 100% (${count} user stories = ${count}+ E2E tests)`);
-    userStories.slice(0, 3).forEach((story, i) => {
-      console.log(`         ${i + 1}. ${story.story_id}: ${story.title} (${story.story_points} pts)`);
-    });
-
-    if (count > 3) {
-      console.log(`         ... and ${count - 3} more`);
-    }
-  }
-
-  return {
-    user_stories_count: count,
-    user_stories: userStories || [],
-    test_coverage_target: '100%',
-    expected_test_count: count
-  };
-}
-
-/**
- * Phase 3: Execute E2E Tests
- */
-async function executeE2ETests(sdId, options) {
-  console.log('   🎭 Executing Playwright E2E tests...');
-
-  const results = {
-    tests_executed: 0,
-    tests_passed: 0,
-    failed_tests: 0,
-    skipped_tests: 0,
-    execution_time_ms: 0,
-    failures: [],
-    report_url: null,
-    troubleshooting_tactics: []
-  };
-
-  try {
-    if (options.full_e2e) {
-      console.log('      🚀 Full E2E suite requested');
-      console.log('      💡 Would execute: npm run test:e2e');
-      console.log('      ⏭️  Simulated execution (implement actual test runner)');
-
-      // In real implementation, would run:
-      // const { stdout, stderr } = await execAsync('npm run test:e2e'); // Run from ehg directory
-      // Parse results from stdout
-
-      // Simulated results
-      results.tests_executed = 10;
-      results.tests_passed = 10;
-      results.execution_time_ms = 15000;
-      results.report_url = `tests/e2e/evidence/${sdId}/playwright-report.html`;
-
-      console.log('      ✅ Simulated: 10/10 tests passed (15s)');
-    } else {
-      console.log('      ℹ️  Full E2E suite not requested (use --full-e2e flag)');
-      console.log('      💡 Checking for existing test evidence...');
-
-      // Check database for previous test results
-      // FIX: Use column-specific select to avoid fetching bloated metadata
-      const { data: previousTest, error } = await supabase
-        .from('sub_agent_execution_results')
-        .select('id, sd_id, sub_agent_code, verdict, confidence, metadata, created_at')
-        .eq('sd_id', sdId)
-        .eq('sub_agent_code', 'TESTING')
-        .order('created_at', { ascending: false })
-        .limit(1)
-        .single();
-
-      if (previousTest && !error) {
-        // Check if cached results are recent (within 1 hour)
-        const testAge = Date.now() - new Date(previousTest.created_at).getTime();
-        const oneHour = 60 * 60 * 1000;
-        const isRecent = testAge < oneHour;
-
-        console.log('      ✅ Found previous test execution');
-        console.log(`         Verdict: ${previousTest.verdict}`);
-        console.log(`         Date: ${previousTest.created_at}`);
-        console.log(`         Age: ${Math.floor(testAge / 1000 / 60)} minutes ago`);
-
-        if (isRecent) {
-          console.log('      ♻️  Using cached results (< 1 hour old)');
-
-          // Extract test execution data from cached result
-          // Try multiple paths: metadata.findings.phase3_execution (new format) or direct metadata fields (old format)
-          // FIX: Add size validation to prevent using bloated cached data
-          const rawCachedData = previousTest.metadata?.findings?.phase3_execution || previousTest.metadata || {};
-          const cachedDataSize = JSON.stringify(rawCachedData).length;
-          const MAX_CACHE_SIZE = 50000; // 50 KB threshold
-
-          let cachedData = rawCachedData;
-          if (cachedDataSize > MAX_CACHE_SIZE) {
-            console.log(`      ⚠️  Cached data too large (${Math.round(cachedDataSize / 1024)} KB > ${MAX_CACHE_SIZE / 1024} KB threshold)`);
-            console.log('      💡 Extracting only essential fields to prevent bloat');
-            // Extract only essential fields, skip potentially bloated nested data
-            cachedData = {
-              tests_executed: rawCachedData.tests_executed || 0,
-              tests_passed: rawCachedData.tests_passed || 0,
-              failed_tests: rawCachedData.failed_tests || 0,
-              _size_limited: true,
-              _original_size_kb: Math.round(cachedDataSize / 1024)
-            };
-          }
-          results.tests_executed = cachedData.tests_executed || 0;
-          results.tests_passed = cachedData.tests_passed || 0;
-          results.failed_tests = cachedData.failed_tests || 0;
-          results.from_cache = true;
-          results.cache_age_minutes = Math.floor(testAge / 1000 / 60);
-
-          // If we have a valid cached verdict but no test data, log warning but accept the verdict
-          if (results.tests_executed === 0 && (previousTest.verdict === 'PASS' || previousTest.verdict === 'CONDITIONAL_PASS')) {
-            console.log('      ⚠️  Cached verdict found but test execution data missing');
-            console.log('      💡 Accepting cached verdict:', previousTest.verdict);
-            // Set minimal valid data to avoid "No E2E tests executed" error
-            results.tests_executed = 1;
-            results.tests_passed = previousTest.verdict === 'PASS' ? 1 : 1;
-            results.failed_tests = previousTest.verdict === 'CONDITIONAL_PASS' ? 1 : 0;
-          }
-        } else {
-          console.log('      ⏰ Cached results too old (> 1 hour)');
-          console.log('      💡 Re-run tests with --full-e2e flag for fresh results');
-        }
-      } else {
-        console.log('      ⚠️  No previous test evidence found');
-        console.log('      💡 Run tests with --full-e2e flag to execute E2E suite');
-      }
-    }
-  } catch (error) {
-    console.error(`      ❌ Test execution error: ${error.message}`);
-    results.error = error.message;
-
-    // Provide troubleshooting guidance based on error type
-    results.troubleshooting_tactics = suggestTroubleshootingTactics(error);
-
-    console.log('\n      🔧 TROUBLESHOOTING SUGGESTIONS:');
-    results.troubleshooting_tactics.forEach((tactic, i) => {
-      console.log(`         ${i + 1}. ${tactic.name} (${tactic.tier})`);
-      console.log(`            ${tactic.description}`);
-    });
-  }
-
-  return results;
-}
-
-/**
- * Suggest troubleshooting tactics based on error type
- */
-function suggestTroubleshootingTactics(error) {
-  const tactics = [];
-  const errorMsg = error.message.toLowerCase();
-
-  // Tactic 1: Always suggest server restart + single test first
-  tactics.push({
-    name: 'Server Kill & Restart + Single Test Isolation',
-    tier: 'Tier 1 (Quick Win)',
-    description: 'Kill server, restart fresh, run single test in isolation',
-    command: process.platform === 'win32' ? 'taskkill /f /im node.exe & npm run dev' : 'pkill -f "vite" && npm run dev',
-    priority: 1,
-    estimated_time: '5-10 minutes',
-    fixes_percentage: '40%'
-  });
-
-  // Port conflict
-  if (errorMsg.includes('eaddrinuse') || errorMsg.includes('address already in use') ||
-      errorMsg.includes('connection refused')) {
-    tactics.push({
-      name: 'Port Conflict Resolution',
-      tier: 'Tier 1 (Quick Win)',
-      description: 'Free up ports blocked by zombie processes',
-      command: process.platform === 'win32'
-        ? 'netstat -ano | findstr :5173  (then: taskkill /F /PID [PID])'
-        : 'lsof -i :5173 && kill -9 [PID]',
-      priority: 1,
-      estimated_time: '5 minutes'
-    });
-  }
-
-  // Module/cache issues
-  if (errorMsg.includes('module') || errorMsg.includes('cannot find') ||
-      errorMsg.includes('enoent')) {
-    tactics.push({
-      name: 'Nuclear Cache Clear',
-      tier: 'Tier 1 (Quick Win)',
-      description: 'Remove all cached build artifacts',
-      command: process.platform === 'win32'
-        ? 'rmdir /s /q node_modules\\.vite dist && npm run build'
-        : 'rm -rf node_modules/.vite dist/ && npm run build',
-      priority: 1,
-      estimated_time: '10-15 minutes'
-    });
-
-    tactics.push({
-      name: 'Dependency Lock Verification',
-      tier: 'Tier 1 (Quick Win)',
-      description: 'Ensure package-lock.json matches installed versions',
-      command: 'npm ci && npx playwright install --with-deps',
-      priority: 2,
-      estimated_time: '10 minutes'
-    });
-  }
-
-  // Timeout issues
-  if (errorMsg.includes('timeout') || errorMsg.includes('exceeded')) {
-    tactics.push({
-      name: 'Test Timeout & Async Analysis',
-      tier: 'Tier 2 (Deep Diagnostic)',
-      description: 'Increase timeout and check for missing await statements',
-      command: 'npx playwright test --timeout=60000 --debug',
-      priority: 2,
-      estimated_time: '15-20 minutes'
-    });
-  }
-
-  // Element not found
-  if (errorMsg.includes('element') || errorMsg.includes('selector') ||
-      errorMsg.includes('not found')) {
-    tactics.push({
-      name: 'Visual Debugging & Screenshots',
-      tier: 'Tier 3 (Advanced)',
-      description: 'See exactly what browser sees at failure point',
-      command: 'npx playwright test --headed --debug',
-      priority: 2,
-      estimated_time: '20-30 minutes'
-    });
-  }
-
-  // Database issues
-  if (errorMsg.includes('database') || errorMsg.includes('query') ||
-      errorMsg.includes('rls') || errorMsg.includes('permission')) {
-    tactics.push({
-      name: 'Database State Verification & Reset',
-      tier: 'Tier 2 (Deep Diagnostic)',
-      description: 'Ensure test database is in known state',
-      command: 'npm run db:migrate:status && npm run db:seed:test',
-      priority: 2,
-      estimated_time: '15-20 minutes'
-    });
-  }
-
-  // Environment issues
-  if (errorMsg.includes('undefined') || errorMsg.includes('env') ||
-      errorMsg.includes('config')) {
-    tactics.push({
-      name: 'Environment Variable Validation',
-      tier: 'Tier 2 (Deep Diagnostic)',
-      description: 'Verify all required env vars are loaded',
-      command: 'node -e "require(\'dotenv\').config(); console.log(process.env.SUPABASE_URL)"',
-      priority: 2,
-      estimated_time: '10 minutes'
-    });
-  }
-
-  // Sort by priority
-  tactics.sort((a, b) => a.priority - b.priority);
-
-  return tactics;
-}
-
-/**
- * Phase 4: Collect Evidence
- */
-async function collectEvidence(sdId, phase3Results) {
-  console.log('   📸 Collecting test evidence...');
-
-  const evidence = {
-    screenshots: [],
-    reports: [],
-    logs: []
-  };
-
-  if (phase3Results.report_url) {
-    evidence.reports.push({
-      type: 'playwright_html',
-      url: phase3Results.report_url,
-      description: 'Playwright HTML test report'
-    });
-    console.log(`      ✅ Report: ${phase3Results.report_url}`);
-  }
-
-  if (phase3Results.tests_executed > 0) {
-    evidence.screenshots.push({
-      count: phase3Results.tests_passed,
-      description: `Screenshots for ${phase3Results.tests_passed} passing tests`
-    });
-    console.log(`      ✅ Screenshots: ${phase3Results.tests_passed} captured`);
-  }
-
-  console.log(`      💾 Evidence stored in: tests/e2e/evidence/${sdId}/`);
-
-  return evidence;
-}
-
-/**
- * Phase 5: Generate Verdict
- */
-function generateVerdict(results, validationMode = 'prospective') {
-  const { findings, critical_issues, warnings } = results;
-
-  let verdict = 'PASS';
-  let confidence = 100;
-  const recommendations = [];
-  let justification = null;  // For CONDITIONAL_PASS
-  let conditions = null;  // For CONDITIONAL_PASS
-
-  // SD-LEO-PROTOCOL-V4-4-0: Adaptive validation mode logging
-  console.log(`   📋 Applying ${validationMode} validation criteria...`);
-
-  // Critical issues = BLOCKED
-  if (critical_issues.length > 0) {
-    verdict = 'BLOCKED';
-    confidence = 100;
-    recommendations.push('Fix all critical issues before proceeding');
-    recommendations.push('📖 Consult Troubleshooting Tactics Arsenal in TESTING sub-agent description');
-  }
-  // Failed tests - check pass rate threshold (95%)
-  else if (findings.phase3_execution?.failed_tests > 0) {
-    const testsExecuted = findings.phase3_execution?.tests_executed || 0;
-    const testsPassed = findings.phase3_execution?.tests_passed || 0;
-    const passRate = testsExecuted > 0 ? (testsPassed / testsExecuted) * 100 : 0;
-    const PASS_RATE_THRESHOLD = 95;
-
-    if (passRate >= PASS_RATE_THRESHOLD) {
-      // High pass rate (≥95%) - mode-dependent verdict
-      // SD-LEO-PROTOCOL-V4-4-0: CONDITIONAL_PASS only allowed in retrospective mode
-      if (validationMode === 'retrospective') {
-        verdict = 'CONDITIONAL_PASS';
-        confidence = 90;
-        justification = `High pass rate (${passRate.toFixed(1)}% >= ${PASS_RATE_THRESHOLD}%) with ${findings.phase3_execution.failed_tests} minor failure(s). Tests demonstrate functional coverage.`;
-        conditions = [
-          `Fix remaining ${findings.phase3_execution.failed_tests} test failure(s) for 100% coverage`,
-          'Review failed test details in phase3_execution.failures'
-        ];
-        recommendations.push(`Pass rate ${passRate.toFixed(1)}% meets threshold (≥${PASS_RATE_THRESHOLD}%)`);
-        recommendations.push(`${findings.phase3_execution.failed_tests} minor test failures - consider fixing for 100% coverage`);
-      } else {
-        // Prospective mode: still PASS since threshold met, but with warning
-        verdict = 'PASS';
-        confidence = 90;
-        recommendations.push(`Pass rate ${passRate.toFixed(1)}% meets threshold (≥${PASS_RATE_THRESHOLD}%)`);
-        recommendations.push(`${findings.phase3_execution.failed_tests} minor test failures - recommend fixing before deployment`);
-      }
-    } else {
-      // Low pass rate (<95%) - block
-      verdict = 'BLOCKED';
-      confidence = 100;
-      recommendations.push(`Pass rate ${passRate.toFixed(1)}% below threshold (${PASS_RATE_THRESHOLD}%)`);
-      recommendations.push('Fix failing E2E tests before approval');
-      recommendations.push('🔧 Use Troubleshooting Arsenal: Start with Tactic 1 (Server Restart + Single Test)');
-
-      // Add specific troubleshooting tactics if available
-      if (findings.phase3_execution?.troubleshooting_tactics?.length > 0) {
-        const topTactic = findings.phase3_execution.troubleshooting_tactics[0];
-        recommendations.push(`💡 Suggested: ${topTactic.name} - ${topTactic.command}`);
-      }
-    }
-  }
-  // Test execution error
-  else if (findings.phase3_execution?.error) {
-    verdict = 'BLOCKED';
-    confidence = 100;
-    recommendations.push('Resolve test execution error before approval');
-    recommendations.push('🔧 Troubleshooting tactics have been suggested in Phase 3 output');
-
-    // Include top 2 troubleshooting tactics in recommendations
-    if (findings.phase3_execution?.troubleshooting_tactics?.length > 0) {
-      findings.phase3_execution.troubleshooting_tactics.slice(0, 2).forEach(tactic => {
-        recommendations.push(`   ${tactic.name}: ${tactic.command}`);
-      });
-    }
-  }
-  // No tests executed - ADAPTIVE LOGIC
-  else if (findings.phase3_execution?.tests_executed === 0) {
-    if (validationMode === 'retrospective') {
-      // SD-LEO-PROTOCOL-V4-4-0: Retrospective mode - accept if evidence of testing exists
-      const testsPassed = findings.phase3_execution?.tests_passed || 0;
-      const testFilesFound = findings.phase4_evidence?.test_files_found || 0;
-      // Also check if user stories have e2e mappings (Phase 4.5 validation)
-      const userStoriesWithE2E = results.critical_issues?.length === 0;  // No critical issues = all stories verified
-      const hasTestEvidence = testsPassed > 0 || testFilesFound > 0 || userStoriesWithE2E;
-
-      if (hasTestEvidence) {
-        verdict = 'CONDITIONAL_PASS';
-        confidence = 75;
-        justification = `E2E testing completed retrospectively. Evidence: ${testsPassed} tests passed. Work already delivered and functional. Missing --full-e2e flag is infrastructure gap, not functional failure.`;
-        conditions = [
-          'Recommend: Add --full-e2e flag to CI/CD pipeline for future SDs',
-          'Consider: Create follow-up SD for testing infrastructure improvements'
-        ];
-        recommendations.push('Accepted retrospectively - work delivered successfully');
-        recommendations.push('Infrastructure gap documented in conditions');
-      } else {
-        // No evidence in retrospective mode = still BLOCKED
-        verdict = 'BLOCKED';
-        confidence = 100;
-        recommendations.push('No E2E test evidence found - cannot validate retrospectively');
-        recommendations.push('Manual validation or test execution required');
-      }
-    } else {
-      // Prospective mode - strict enforcement
-      verdict = 'BLOCKED';
-      confidence = 100;
-      recommendations.push('Execute E2E tests before approval (MANDATORY - zero tolerance)');
-      recommendations.push('E2E testing is NOT optional per protocol - all tests must pass with zero failures');
-      recommendations.push('Use: node scripts/execute-subagent.js --code TESTING --sd-id <SD-ID> --full-e2e');
-    }
-  }
-  // Warnings present = mode-dependent verdict
-  // SD-LEO-PROTOCOL-V4-4-0: CONDITIONAL_PASS only allowed in retrospective mode
-  else if (warnings.length > 0) {
-    if (validationMode === 'retrospective') {
-      // Retrospective: CONDITIONAL_PASS acceptable for completed work with minor issues
-      verdict = 'CONDITIONAL_PASS';
-      confidence = 85;
-      justification = `Testing completed retrospectively with ${warnings.length} warning(s). Work delivered successfully but minor issues documented for future improvement.`;
-      conditions = warnings.map(w => `Address: ${w.issue || w}`).slice(0, 5);  // Limit to 5 conditions
-      recommendations.push('Address warnings for improved quality');
-    } else {
-      // Prospective: Cannot use CONDITIONAL_PASS, use WARNING verdict instead
-      verdict = 'PASS';  // Tests passed, warnings are non-blocking
-      confidence = 85;
-      recommendations.push(`${warnings.length} warning(s) found - address for improved quality`);
-      recommendations.push('Re-run with --validation-mode retrospective if work is already complete');
-    }
-  }
-  // All passed = PASS
-  else {
-    verdict = 'PASS';
-    confidence = 95;
-    recommendations.push('All tests passed - ready for deployment');
-  }
-
-  // Additional recommendations
-  if (findings.phase2_test_generation?.user_stories_count === 0) {
-    recommendations.push('Create user stories to enable comprehensive E2E test coverage');
-  }
-
-  if (!findings.phase3_execution?.from_cache) {
-    recommendations.push('Test evidence is fresh (not cached)');
-  }
-
-  // Add troubleshooting arsenal reference if there were any issues
-  if (verdict === 'BLOCKED' || verdict === 'CONDITIONAL_PASS') {
-    recommendations.push('📚 Full troubleshooting arsenal (13 tactics) available in TESTING sub-agent description');
-    recommendations.push('⏱️  Expected debugging time savings: 3-8x with systematic troubleshooting');
-  }
-
-  return {
-    verdict,
-    confidence,
-    recommendations,
-    justification,  // SD-LEO-PROTOCOL-V4-4-0: For CONDITIONAL_PASS
-    conditions,  // SD-LEO-PROTOCOL-V4-4-0: For CONDITIONAL_PASS follow-up actions
-    summary: `${findings.phase3_execution?.tests_passed || 0}/${findings.phase3_execution?.tests_executed || 0} tests passed`,
-    troubleshooting_available: true
-  };
-}
+// Re-export from modular structure
+export { execute } from './testing/index.js';
