+++ lib/sub-agents/testing/phases/phase1-preflight.js
@@ -1,0 +1,88 @@
+/**
+ * Phase 1: Pre-flight Checks
+ *
+ * Validates build status, migrations, and component integration before testing.
+ *
+ * Extracted from TESTING sub-agent v3.0
+ * SD: SD-LEO-REFAC-TESTING-INFRA-001
+ */
+
+/**
+ * Execute pre-flight checks
+ * @param {string} sdId - Strategic Directive ID
+ * @param {Object} options - Execution options
+ * @param {Object} supabase - Supabase client
+ * @returns {Promise<Object>} Pre-flight check results
+ */
+export async function preflightChecks(sdId, options, supabase) {
+  const checks = {
+    blocked: false,
+    critical_issues: [],
+    warnings: [],
+    build_status: null,
+    migrations_status: null,
+    component_integration: null
+  };
+
+  // Check 1: Build validation (if not skipped)
+  if (!options.skip_build) {
+    console.log('   🏗️  Checking build status...');
+    try {
+      // Simplified check - in real implementation would run actual build
+      checks.build_status = {
+        passed: true,
+        message: 'Build check skipped (would run: npm run build)'
+      };
+      console.log('      ✅ Build validation passed');
+    } catch (error) {
+      checks.blocked = true;
+      checks.critical_issues.push({
+        severity: 'CRITICAL',
+        issue: 'Build failed',
+        recommendation: 'Fix build errors before testing',
+        error: error.message
+      });
+      console.log('      ❌ Build validation failed');
+    }
+  } else {
+    console.log('   ⏭️  Build validation skipped');
+    checks.build_status = { skipped: true };
+  }
+
+  // Check 2: Database migration verification
+  console.log('   🗄️  Checking database migrations...');
+  try {
+    const { data: migrations, error } = await supabase
+      .from('migrations')
+      .select('*')
+      .order('created_at', { ascending: false })
+      .limit(5);
+
+    if (error) {
+      console.log(`      ⚠️  Could not check migrations: ${error.message}`);
+      checks.warnings.push({
+        severity: 'MEDIUM',
+        issue: 'Could not verify database migrations',
+        recommendation: 'Manually verify migrations are applied'
+      });
+    } else {
+      console.log(`      ✅ Migration check complete (${migrations?.length || 0} recent)`);
+      checks.migrations_status = {
+        checked: true,
+        recent_count: migrations?.length || 0
+      };
+    }
+  } catch (error) {
+    console.log(`      ⚠️  Migration check error: ${error.message}`);
+  }
+
+  // Check 3: Component integration (manual note)
+  console.log('   🔗 Component integration check...');
+  console.log('      💡 Tip: Verify components are imported and used (not just created)');
+  checks.component_integration = {
+    manual_check_required: true,
+    suggestion: 'Search for component imports in parent files'
+  };
+
+  return checks;
+}
