+++ lib/sub-agents/testing/phases/phase2-generation.js
@@ -1,0 +1,54 @@
+/**
+ * Phase 2: Test Case Generation
+ *
+ * Generates test cases from user stories for the SD.
+ *
+ * Extracted from TESTING sub-agent v3.0
+ * SD: SD-LEO-REFAC-TESTING-INFRA-001
+ */
+
+/**
+ * Generate test cases from user stories
+ * @param {string} sdId - Strategic Directive ID
+ * @param {Object} _options - Execution options (unused currently)
+ * @param {Object} supabase - Supabase client
+ * @returns {Promise<Object>} Test case generation results
+ */
+export async function generateTestCases(sdId, _options, supabase) {
+  console.log('   📋 Querying user stories...');
+
+  const { data: userStories, error } = await supabase
+    .from('user_stories')
+    .select('*')
+    .eq('sd_id', sdId)
+    .order('story_points', { ascending: false });
+
+  if (error) {
+    console.log(`      ⚠️  Could not query user stories: ${error.message}`);
+    return {
+      user_stories_count: 0,
+      error: error.message
+    };
+  }
+
+  const count = userStories?.length || 0;
+  console.log(`      ✅ Found ${count} user stories`);
+
+  if (count > 0) {
+    console.log(`      💡 Test Coverage Target: 100% (${count} user stories = ${count}+ E2E tests)`);
+    userStories.slice(0, 3).forEach((story, i) => {
+      console.log(`         ${i + 1}. ${story.story_id}: ${story.title} (${story.story_points} pts)`);
+    });
+
+    if (count > 3) {
+      console.log(`         ... and ${count - 3} more`);
+    }
+  }
+
+  return {
+    user_stories_count: count,
+    user_stories: userStories || [],
+    test_coverage_target: '100%',
+    expected_test_count: count
+  };
+}
